Refactor: Move components to own files from index.tsx
