+++ src/components/frame/v5/pages/UserCryptoToFiatPage/partials/BankDetailsModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BankDetailsModal.tsx';
