+++ src/components/frame/v5/pages/UserCryptoToFiatPage/partials/ContactDetailsForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ContactDetailsForm.tsx';
